+++ client/web/src/components/MapPage/layouts/Desktop.js
