+++ client/web/src/components/MapPage/layouts/Mobile.js
